For LiveTV, prefer the first available tuner on the same host as the frontend. Else, the first available tuner anywhere.
git-svn-id: http://svn.mythtv.org/svn/trunk@1457 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmyth/util.cpp
@@ -134,7 +134,7 @@ void WriteBlock(QSocket *socket, void *data, int len)
         socket->flush();
     qApp->unlock();
 
-    while (socket->bytesToWrite() >= written)
+    while (socket->bytesToWrite() >= (unsigned) written)
         usleep(50000);
 }
